requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0
jinja2>=3.1
python-multipart>=0.0.9
httpx>=0.27
pydantic-settings>=2.0
itsdangerous>=2.1
